Most of invert done
